Fix - Move categories to its own function file
